renamed folder containing custom CMS tags to be more emblematic
